attempt to fix specific model testing

+++ src/test/java/com/jtrull/alzdetection/TestImagePrediction.java
@@ -1,5 +1,6 @@
 package com.jtrull.alzdetection;
 
+import org.junit.Assert;
 import org.junit.experimental.ParallelComputer;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -9,6 +10,8 @@ import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.runner.JUnitCore;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -37,6 +40,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileWriter;
 import java.io.InputStream;
@@ -52,6 +56,7 @@ import java.util.Random;
 @AutoConfigureMockMvc
 @TestMethodOrder(OrderAnnotation.class)
 public class TestImagePrediction {
+    Logger logger = LoggerFactory.getLogger(TestImagePrediction.class);
     @Autowired
 	private MockMvc mvc;
 
@@ -66,15 +71,20 @@ public class TestImagePrediction {
 
     private static final String BASE_URL = "/api/v1/model";
 
+    private int num = 0;
+    private static final int TEST_INVOCATIONS = 1;
+
     private String createBaseUrl(Long modelId) {
         return BASE_URL + "/" + String.valueOf(modelId) + "/predict";
     }
 
     @Test
 	@Order(1)
-    @RepeatedTest(10)
+    @RepeatedTest(TEST_INVOCATIONS)
 	public void testRandomPrediction() throws Exception {
-        MvcResult _return = mvc.perform(get(createBaseUrl(getModel().getId()) + "/random")
+        String url = createBaseUrl(getModel().getId());
+        logger.info("Running random prediction on url: " + url);
+        MvcResult _return = mvc.perform(get(url + "/random")
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().is2xxSuccessful())
             .andReturn();
@@ -89,7 +99,7 @@ public class TestImagePrediction {
 
     @Test
 	@Order(1)
-    @RepeatedTest(10)
+    @RepeatedTest(TEST_INVOCATIONS)
 	public void testRandomPredictionFromCategory() throws Exception {
         for (ImpairmentEnum val : ImpairmentEnum.values()) {
             MvcResult _return = mvc.perform(get(createBaseUrl(getModel().getId()) + "/" + val.toString())
@@ -108,26 +118,28 @@ public class TestImagePrediction {
 
     @Test
 	@Order(1)
-    @RepeatedTest(10)
+    @RepeatedTest(TEST_INVOCATIONS)
 	public void testRandomPredictionFromInvalidCategory() throws Exception {
         String impairment = RandomStringUtils.random(5, true, true);
 
         try {
-            mvc.perform(get(createBaseUrl(getModel().getId()) + "/" + impairment)
+            mvc.perform(get(createBaseUrl(getModel(1L).getId()) + "/" + impairment)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().is4xxClientError());
             fail("succeeded sending invalid impairment category when expected to fail");
 
         } catch (ServletException e) {
             RestClientResponseException httpException = (RestClientResponseException) e.getRootCause();
-            assert httpException.getStatusCode().equals(HttpStatus.valueOf(400));
-            assert httpException.getMessage().contains( "Unable to parse category: " + impairment + ". Expected values=[" + Arrays.asList(ImpairmentEnum.asStrings().toArray()) + "]");
+            Assert.assertTrue("status code did not match 400 as expected, found: " + httpException.getStatusCode(), 
+                httpException.getStatusCode().equals(HttpStatus.valueOf(400)));
+            Assert.assertTrue("status message was not as expected, found: " + httpException.getStatusCode(), 
+                httpException.getMessage().contains( "Unable to parse category: " + impairment + ". Expected values=[" + Arrays.asList(ImpairmentEnum.asStrings().toArray()) + "]"));
         }
 	}
 
     @Test
 	@Order(2)
-    @RepeatedTest(10)
+    @RepeatedTest(TEST_INVOCATIONS)
 	public void testPredictionFromFile() throws Exception {
         List<ImagePrediction> allImages = imageRepository.findAll();
         if (allImages.size() == 0) {
@@ -171,26 +183,28 @@ public class TestImagePrediction {
         assert prediction.getConf_ModerateImpairment().equals(initialImagePrediction.getConf_ModerateImpairment());
 	}
 
+    
     @Test
 	@Order(2)
-    @RepeatedTest(10)
+    @RepeatedTest(TEST_INVOCATIONS)
     public void testPredictionFromInvalidFile() throws Exception {
         String path = imageService.returnImagePath();
-		String filepath = path + "/test.json";
-
-		JSONObject json = new JSONObject();
-		json.put("i am a json file", "not a jpg");
-		FileWriter file = new FileWriter(filepath);
-		file.write(json.toString());
-		file.close();
-		FileInputStream fis = new FileInputStream(filepath);
-
+		String filepath = path + "/test-" + num++ + ".json";
+        File f = new File(filepath);
+        if (!f.exists()) {
+            JSONObject json = new JSONObject();
+            json.put("i am a json file", "not a jpg");
+            FileWriter file = new FileWriter(filepath);
+            file.write(json.toString());
+            file.close();
+        }
+		
 		try (InputStream is = getClass().getResourceAsStream(filepath)) {
 			MockMultipartFile mockMultipartFile = new MockMultipartFile("image", "test.json", 
-            MediaType.APPLICATION_JSON.toString(), ByteStreams.toByteArray(fis));
+            MediaType.APPLICATION_JSON.toString(), ByteStreams.toByteArray(new FileInputStream(filepath)));
 
 			try {
-				mvc.perform(MockMvcRequestBuilders.multipart(createBaseUrl(getModel().getId()))
+				mvc.perform(MockMvcRequestBuilders.multipart(createBaseUrl(getModel(1L).getId()))
 						.file(mockMultipartFile)
 						.contentType(MediaType.APPLICATION_JSON.toString()))
 						.andExpect(status().is4xxClientError());
@@ -198,15 +212,36 @@ public class TestImagePrediction {
 
 			} catch (ServletException e) {
 				RestClientResponseException httpException = (RestClientResponseException) e.getRootCause();
-                assert httpException.getStatusCode().equals(HttpStatus.valueOf(400));
-				assert httpException.getMessage().contains("is it a valid image?");
+                Assert.assertTrue("status message was not as expected, found: " + httpException.getStatusCode(), 
+                    httpException.getMessage().contains("is it a valid image?"));
+                Assert.assertTrue("status code did not match 400 as expected, found: " + httpException.getStatusCode(), 
+                    httpException.getStatusCode().equals(HttpStatus.valueOf(400)));
 			}
 		}
     }
 
+    
+
+    /**
+	 * Run all tests in class concurrently
+	 */
+	@Test
+	@Order(4)
+	@RepeatedTest(TEST_INVOCATIONS)
+    public void runAllTests() {
+        int numConcurrent = 25_000;
+        Class<?>[] classes  = new Class<?>[numConcurrent];
+        Arrays.fill(classes, TestImagePrediction.class);
+        JUnitCore.runClasses(new ParallelComputer(true, true), classes);
+    }
+
+    public Model getModel() throws Exception {
+        return getModel(null);
+    }
+
     @Test
-    @Order(3)
-    @RepeatedTest(10)
+    @Order(5)
+    @RepeatedTest(TEST_INVOCATIONS)
     public void testDeletePrediction() throws Exception {
         Optional<ImagePrediction> imageOpt = imageRepository.findAll().stream().findFirst();
         if (imageOpt.isEmpty()) {
@@ -214,7 +249,7 @@ public class TestImagePrediction {
         }
         ImagePrediction initialImagePrediction = imageOpt.get();
 
-        MvcResult _return = mvc.perform(delete(createBaseUrl(getModel().getId()) + "/" + initialImagePrediction.getId())
+        MvcResult _return = mvc.perform(delete(createBaseUrl(getModel(initialImagePrediction.getAssociatedModel()).getId()) + "/" + initialImagePrediction.getId())
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().is2xxSuccessful())
             .andReturn();
@@ -226,8 +261,8 @@ public class TestImagePrediction {
     }
 
     @Test
-    @Order(3)
-    @RepeatedTest(10)
+    @Order(5)
+    @RepeatedTest(TEST_INVOCATIONS)  
     public void testInvalidDeletePrediction() throws Exception {
         long invalidId = 2345234523452345L;
         try {
@@ -238,22 +273,11 @@ public class TestImagePrediction {
 
         } catch (ServletException e) {
             RestClientResponseException httpException = (RestClientResponseException) e.getRootCause();
-            assert httpException.getStatusCode().equals(HttpStatus.valueOf(404));
-            assert httpException.getMessage().contains("Unable to find prediction with specified Id: " + invalidId);
+            Assert.assertTrue("status code did not match 404 as expected, found: " + httpException.getStatusCode(), 
+                httpException.getStatusCode().equals(HttpStatus.valueOf(404)));
+            Assert.assertTrue("status message was not as expected, found: " + httpException.getStatusCode(), 
+                httpException.getMessage().contains("Unable to find prediction with specified Id: " + invalidId));
         }
-    }
-
-    /**
-	 * Run all tests in class concurrently
-	 */
-	@Test
-	@Order(5)
-	@RepeatedTest(10)
-    public void runAllTests() {
-        int numConcurrent = 25_000;
-        Class<?>[] classes  = new Class<?>[numConcurrent];
-        Arrays.fill(classes, TestImagePrediction.class);
-        JUnitCore.runClasses(new ParallelComputer(true, true), classes);
     }
 
     /**
@@ -262,23 +286,27 @@ public class TestImagePrediction {
      * @return
      * @throws Exception
      */
-    public Model getModel() throws Exception {
-        List<Model> allModels = modelRepository.findAll();
+    public Model getModel(Long desiredId) throws Exception {
+        synchronized (modelRepository) {
+            // List<Model> allModels = modelRepository.findAll();
+            // logger.info("found " + allModels.size() + " models in repo");
+            // // load models into the model repository
+            // for (int i = allModels.size(); i<=10; i++) {
+            //     TestModel.runLoadModelRequest(modelService, getClass(), mvc);
+            // }
+            // allModels = modelRepository.findAll();
+            // logger.info("after push, found " + allModels.size() + " models in repo");
 
-        // load 10 models into the model repository
-        long modelNum = 0L;
-        for (int i=allModels.size(); i<=10; i++) {
-            TestModel.runLoadModelRequest(modelService, modelNum++, getClass(), mvc);
+            // if (desiredId != null) {
+            //     return allModels.stream().filter(m -> m.getId() == desiredId).findFirst().get();
+            // }
+
+            // // pick a random model
+            // return allModels.stream()
+            //     .skip(new Random().nextInt(allModels.size()/5))
+            //     .findAny()
+            //     .orElseThrow(() -> new RuntimeException("Unable to pick random model"));
+            return modelRepository.findAll().stream().findFirst().get();
         }
-        allModels = modelRepository.findAll();
-
-        // pick a random model
-        Optional<Model> randomModelOpt = allModels.stream()
-            .skip(new Random().nextInt(allModels.size()))
-            .findAny();
-
-        if (randomModelOpt.isPresent()) return randomModelOpt.get();
-
-        throw new RuntimeException("Unable to find model");
     }
 }
